+++ examples/text-generator/encoder.py
@@ -115,16 +115,3 @@ def get_encoder():
     )
     bpe_merges = [tuple(merge_str.split()) for merge_str in bpe_data.split("\n")[1:-1]]
     return Encoder(encoder=encoder, bpe_merges=bpe_merges)
-
-
-encoder = get_encoder()
-
-
-def pre_inference(sample, metadata):
-    context = encoder.encode(sample["text"])
-    return {"context": [context]}
-
-
-def post_inference(prediction, metadata):
-    response = prediction["response"]["sample"]
-    return {encoder.decode(response)}
