+++ pkg/operator/workloads/api_workload.go
@@ -297,7 +297,7 @@ func tfAPISpec(
 						ImagePullPolicy: "Always",
 						Args: []string{
 							"--download_from=" + config.AWS.S3Path(ctx.ProjectKey),
-							"--download_to=" + path.Join(consts.EmptyDirMountPath, "project", "project.zip"),
+							"--download_to=" + path.Join(consts.EmptyDirMountPath, "project"),
 						},
 						Env:          k8s.AWSCredentials(),
 						VolumeMounts: k8s.DefaultVolumeMounts(),
@@ -316,6 +316,7 @@ func tfAPISpec(
 							"--api=" + ctx.APIs[api.Name].ID,
 							"--model-dir=" + path.Join(consts.EmptyDirMountPath, "model"),
 							"--cache-dir=" + consts.ContextCacheDir,
+							"--project-dir=" + path.Join(consts.EmptyDirMountPath, "project"),
 						},
 						Env:          k8s.AWSCredentials(),
 						VolumeMounts: k8s.DefaultVolumeMounts(),
@@ -460,6 +461,7 @@ func onnxAPISpec(
 							"--api=" + ctx.APIs[api.Name].ID,
 							"--model-dir=" + path.Join(consts.EmptyDirMountPath, "model"),
 							"--cache-dir=" + consts.ContextCacheDir,
+							"--project-dir=" + path.Join(consts.EmptyDirMountPath, "project"),
 						},
 						Env:          k8s.AWSCredentials(),
 						VolumeMounts: k8s.DefaultVolumeMounts(),
